backend/requirements.txt
Flask
Flask-Cors
Flask-Session
SQLAlchemy
bcrypt
PyJWT
python-dotenv
numpy
scipy
pytest
scikit-learn
pylint
requests
email-validator
gunicorn